change file structure to be aligned with angular cli project
