--- src/app.component.ts
@@ -1,25 +1,0 @@
-import { NgModule } from '@angular/core';
-import { BrowserModule } from '@angular/platform-browser';
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'App',
-  template:
-  `<div>
-    <h2>Welcome to {{name}} Angular2!</h2>
-  </div>`
-})
-export class AppComponent implements OnInit {
-  public readonly name = 'electron-forge';
-
-  ngOnInit(): void {
-    console.log('component initialized');
-  }
-}
-
-@NgModule({
-  imports: [BrowserModule],
-  declarations: [AppComponent],
-  bootstrap: [AppComponent]
-})
-export class AppModule { }
