+++ src/app/app.component.html
@@ -1,0 +1,3 @@
+<div>
+    Hello World! {{name}}
+</div>
