+++ src/app/app.component.ts
@@ -1,0 +1,16 @@
+
+
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'App',
+  templateUrl:'./app/app.component.html'
+})
+export class AppComponent implements OnInit {
+  public readonly name = 'electron-forge';
+
+  ngOnInit(): void {
+    console.log('component initialized');
+  }
+}
+
