+++ src/bootstrap.ts
@@ -1,4 +1,4 @@
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
-import { AppModule }  from './app.component';
+import { AppModule }  from './app/app.module';
 
 platformBrowserDynamic().bootstrapModule(AppModule);
